% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dont_run.R
\name{dont_run_get}
\alias{dont_run_get}
\title{Extracts all "don't run" examples in a package}
\usage{
dont_run_get(pkg)
}
\arguments{
\item{pkg}{A \code{pkgdown} package object, or a character path to the location of
the package to parse.}
}
\description{
Extracts all "don't run" examples in a package
}
