% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dont_run.R
\name{dont_run_test}
\alias{dont_run_test}
\title{Copies "don't run" examples from package into individual test scripts and
executes them via the RStudio Job pane.}
\usage{
dont_run_test(pkg, test_dir = NULL)
}
\arguments{
\item{pkg}{A \code{pkgdown} package object, or a character path to the location of
the package to parse.}

\item{test_dir}{Path to where to save the temp test files}
}
\description{
Copies "don't run" examples from package into individual test scripts and
executes them via the RStudio Job pane.
}
