% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autolink.R
\name{ecodown_autolink}
\alias{ecodown_autolink}
\title{Autolink function calls}
\usage{
ecodown_autolink(
  quarto_folder = here::here(),
  render_folder = quarto_entry(quarto_folder, "project", "output-dir"),
  verbosity = c("summary", "verbose", "silent")
)
}
\arguments{
\item{quarto_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{render_folder}{Location of the sub-folder that contains the output
from Quarto. It defaults to the 'output-dir' entry in the '_quarto.yml' file.}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.  Defaults to: 'verbose'}
}
\description{
Autolink function calls
}
