% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{ecodown_clone}
\alias{ecodown_clone}
\title{Download the package's latest source code from repo}
\usage{
ecodown_clone(
  repo_url = "",
  target_folder = tempdir(),
  branch = "main",
  verbosity = c("verbose", "summary", "silent")
)
}
\arguments{
\item{repo_url}{Repo location}

\item{target_folder}{Location to copy the package to. Defaults to a temporary
directory}

\item{branch}{Repo branch. Defaults to 'main'}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.}
}
\description{
Download the package's latest source code from repo
}
\examples{
pkg_path <- ecodown_clone("https://github.com/tidyverse/hms")
list.files(pkg_path)
}
