% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ecodown_convert}
\alias{ecodown_convert}
\title{Copies and/or converts files from package source into Quarto}
\usage{
ecodown_convert(
  package_source_folder = "",
  quarto_sub_folder = "",
  quarto_folder = here::here(),
  convert_readme = TRUE,
  convert_news = TRUE,
  convert_articles = TRUE,
  convert_reference = TRUE,
  downlit_options = TRUE,
  site_url = qe(quarto_folder, "site", "site-url"),
  verbosity = c("verbose", "summary", "silent")
)
}
\arguments{
\item{package_source_folder}{Path to the package's source code}

\item{quarto_sub_folder}{Sub folder in \code{quarto_folder} that will be the base for
the package's documentation.}

\item{quarto_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{convert_readme}{Flag that indicates if the README file needs to be processed}

\item{convert_news}{Flag that indicates if the NEWS file needs to be processed}

\item{convert_articles}{Flag that indicates if the vignette files needs to be processed}

\item{convert_reference}{Flag that indicates if the help files needs to be processed}

\item{downlit_options}{Flag that indicates if the package name should be
added to the 'options()' that tells 'downlit' that this is an internal
package}

\item{site_url}{URL of the target site.  It defaults to using the address in
the '_quarto.yml' file}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.}
}
\description{
Copies and/or converts files from package source into Quarto
}
