% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-versions.R
\name{ecodown_convert_versions}
\alias{ecodown_convert_versions}
\title{Copies and/or converts files from package source into Quarto}
\usage{
ecodown_convert_versions(
  package_source_folder = "",
  package_name = fs::path_file(package_source_folder),
  quarto_sub_folder = package_name,
  version_folder = "",
  quarto_folder = here::here(),
  downlit_options = FALSE,
  site_url = qe(quarto_folder, "site", "site-url"),
  verbosity = c("verbose", "summary", "silent"),
  convert_readme = TRUE,
  convert_news = TRUE,
  convert_articles = TRUE,
  convert_reference = TRUE,
  reference_folder = "reference",
  vignettes_folder = "articles",
  reference_examples = TRUE,
  reference_output = "qmd",
  reference_qmd_options = NULL,
  reference_template = NULL,
  branch = "main",
  versions = list()
)
}
\arguments{
\item{package_source_folder}{Path to the package's source code}

\item{package_name}{Name of the package. Defaults to the top folder
in the repo URL}

\item{quarto_sub_folder}{Sub folder in \code{quarto_folder} that will be the base for
the package's documentation.}

\item{version_folder}{Folder path to save the documentation version. Examples
are: "latest", "dev", "v1.0". Defaults to empty.}

\item{quarto_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{downlit_options}{Flag that indicates if the package name should be
added to the 'options()' that tells 'downlit' that this is an internal
package}

\item{site_url}{URL of the target site.  It defaults to using the address in
the '_quarto.yml' file}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.}

\item{convert_readme}{Flag that indicates if the README file needs to be processed}

\item{convert_news}{Flag that indicates if the NEWS file needs to be processed}

\item{convert_articles}{Flag that indicates if the vignette files needs to be processed}

\item{convert_reference}{Flag that indicates if the help files needs to be processed}

\item{reference_folder}{The sub folder where the reference files will be placed.
Defaults to "reference".}

\item{vignettes_folder}{The sub folder where the vignette files will be placed.
Defaults to "articles".}

\item{reference_examples}{Boolean flag to indicate if the Examples inside the
reference page is to be evaluated.}

\item{reference_output}{File type for all the reference files. Possible options
are \code{qmd} and \code{md}. Defaults to \code{qmd}.}

\item{reference_qmd_options}{A character variable that contains the text version
of additions to the reference front matter. It applies only to when reference
output is \code{qmd}}

\item{reference_template}{The path to a \code{.qmd} file to use as the template to
create the reference pages.}

\item{branch}{Repo branch. Defaults to 'main'}

\item{versions}{A list of additional reference versions to convert.
Pass the commit to use, the name of the version, and the sub-folder to use. It
expects a named list object. Example:
list(list(commit = "latest_commit", path = "dev"))}
}
\description{
Copies and/or converts files from package source into Quarto
}
