% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{examples_run}
\alias{examples_run}
\alias{examples_not_run}
\title{Functions that flag if the example is to run}
\usage{
examples_run()

examples_not_run()
}
\description{
Functions that flag if the example is to run
}
\details{
These functions are placed in the converted \code{.qmd} files
by \code{ecodown_convert()}. They serve as flexible flags that indicate if the
examples, and examples don't run are executed when rendering Quarto.
}
