% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_articles}
\alias{package_articles}
\title{Copies the vignettes into Quarto}
\usage{
package_articles(
  package_source_folder = "",
  source = "vignettes",
  target = "articles",
  quarto_sub_folder = "",
  quarto_base_folder = here::here()
)
}
\arguments{
\item{package_source_folder}{Path to the package's source code}

\item{source}{Sub-folder location where the vignettes are available}

\item{target}{Sub-folder location to place the articles in Quarto}

\item{quarto_sub_folder}{Sub folder in \code{quarto_base_folder} that will be the base for
the package's documentation.}

\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}
}
\description{
Copies the vignettes into Quarto
}
