% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_build_documentation}
\alias{package_build_documentation}
\title{Copies and/or converts files from package source into Quarto}
\usage{
package_build_documentation(
  pkg_folder = "",
  project_folder = "",
  root_folder = here::here(),
  readme = TRUE,
  news = TRUE,
  articles = TRUE,
  reference = TRUE
)
}
\arguments{
\item{pkg_folder}{Path to the package's source code}

\item{project_folder}{Sub folder in \code{root_folder} that will be the base for
the package's documentation.}

\item{root_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{readme}{Flag that indicates of the README file needs to be processed}

\item{news}{Flag that indicates of the NEWS file needs to be processed}

\item{articles}{Flag that indicates of the vignette files needs to be processed}

\item{reference}{Flag that indicates of the help files needs to be processed}
}
\description{
Copies and/or converts files from package source into Quarto
}
