% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_git.R
\name{package_clone_git_repo}
\alias{package_clone_git_repo}
\title{Download the package's latest source code from repo}
\usage{
package_clone_git_repo(
  url = "",
  commit = c("latest_tag", "latest_commit"),
  target_folder = tempdir(),
  branch = "main"
)
}
\arguments{
\item{url}{Repo location}

\item{commit}{Commit to use as the base for the documentation.  It defaults
to 'latest_tag'. That default will search for the latest Git tag.  The
assumption is that the latest tag is the same as the latest release.  This
way we avoid documenting work-in-progress.  The 'latest_commit' value will
simply use whatever is cloned. Pass an SHA value if you wish to fix the
commit to use.}

\item{target_folder}{Location to copy the package to. Defaults to a temporary
directory}

\item{branch}{Repo branch. Defaults to 'main'}
}
\description{
Download the package's latest source code from repo
}
