% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_files.R
\name{package_readme}
\alias{package_readme}
\alias{package_news}
\alias{package_file_copy}
\title{Copies individual files from package to Quarto}
\usage{
package_readme(
  package_source_folder = "",
  target = "",
  file_names = c("README.md"),
  quarto_sub_folder = "",
  quarto_base_folder = here::here()
)

package_news(
  package_source_folder = "",
  target = "",
  file_names = c("NEWS.md", "NEWS.Rmd"),
  quarto_sub_folder = "",
  quarto_base_folder = here::here()
)

package_file_copy(
  package_source_folder = "",
  target = "quarto_sub_folder",
  file_names = c("name.md", "name.Rmd"),
  override_name = NULL,
  quarto_sub_folder = "",
  quarto_base_folder = here::here()
)
}
\arguments{
\item{package_source_folder}{Path to the package's source code}

\item{target}{Sub folder in \code{quarto_sub_folder} where the output file will land}

\item{file_names}{A vector of one or many file names. If it is multiple files
the function will look for and use the first available.}

\item{quarto_sub_folder}{Sub folder in \code{quarto_base_folder} that will be the base for
the package's documentation.}

\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{override_name}{File name of the output, in case the original file name
needs to be changed.}
}
\description{
Copies individual files from package to Quarto
}
