% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_reference.R
\name{package_reference}
\alias{package_reference}
\alias{package_reference_index}
\alias{package_reference_pages}
\title{Processes reference/help files}
\usage{
package_reference(
  pkg_folder = "",
  root_folder = here::here(),
  project_folder = "",
  reference_folder = "reference",
  downlit_options = TRUE,
  url_prefix = ""
)

package_reference_index(
  pkg_folder = "",
  reference_folder = "reference",
  project_folder = "",
  root_folder = here::here(),
  pkg = NULL,
  downlit_options = TRUE,
  url_prefix = ""
)

package_reference_pages(
  pkg_folder = "",
  reference_folder = "reference",
  project_folder = "",
  root_folder = here::here(),
  pkg = NULL,
  downlit_options = TRUE,
  url_prefix = ""
)
}
\arguments{
\item{pkg_folder}{Path to the package's source code}

\item{root_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{project_folder}{Sub folder in \code{root_folder} that will be the base for
the package's documentation.}

\item{reference_folder}{Sub folder in \code{project_folder} where the output files
will land.}

\item{downlit_options}{Flag that indicates if the package name should be
added to the 'options()' that tells 'downlit' that this is an internal
package}

\item{url_prefix}{String to prefix to the 'downlit' URL's}

\item{pkg}{A \code{pkgdown} object. If one is passed, then the \code{pkg_folder} argument
will be ignored}
}
\description{
Processes reference/help files
}
