% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/package_ref_index.R,
%   R/package_ref_pages.R
\name{package_reference}
\alias{package_reference}
\alias{package_reference_index}
\alias{package_reference_pages}
\title{Processes reference/help files}
\usage{
package_reference(
  pkg_folder = "",
  root_folder = here::here(),
  project_folder = "",
  reference_folder = "reference"
)

package_reference_index(
  pkg_folder = "",
  reference_folder = "reference",
  project_folder = "",
  root_folder = here::here(),
  pkg = NULL
)

package_reference_pages(
  pkg_folder = "",
  reference_folder = "reference",
  project_folder = "",
  root_folder = here::here(),
  pkg = NULL
)
}
\arguments{
\item{pkg_folder}{Path to the package's source code}

\item{root_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{project_folder}{Sub folder in \code{root_folder} that will be the base for
the package's documentation.}

\item{reference_folder}{Sub folder in \code{project_folder} where the output files
will land.}

\item{pkg}{A \code{pkgdown} object. If one is passed, then the \code{pkg_folder} argument
will be ignored}
}
\description{
Processes reference/help files
}
