% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qe.R
\name{qe}
\alias{qe}
\title{Get entry from the Quarto setup file}
\usage{
qe(quarto_path = here::here(), ..., file_name = "_quarto.yml")
}
\arguments{
\item{quarto_path}{Path to the Quarto folder, defaults to project path.}

\item{...}{A single or multiple entries with the path to the desired entry.}

\item{file_name}{The name of the Quarto setup file, '_quarto.yml'}
}
\description{
Get entry from the Quarto setup file
}
