% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{site_autolink_html}
\alias{site_autolink_html}
\title{Autolink function calls}
\usage{
site_autolink_html(
  rendered_site = get_quarto_entry(here::here(), "project", "output-dir")
)
}
\arguments{
\item{rendered_site}{Location of the sub-folder that contains the output
from Quarto. It defaults to the 'output-dir' entry in the '_quarto.yml' file.}
}
\description{
Autolink function calls
}
