% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_build.R
\name{site_build_quarto}
\alias{site_build_quarto}
\title{Prepare full Quarto site}
\usage{
site_build_quarto(quarto_base_folder = here::here())
}
\arguments{
\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}
}
\description{
Prepare full Quarto site
}
