% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_quarto.R
\name{site_quarto_build}
\alias{site_quarto_build}
\title{Wrapper on the Quarto Render function}
\usage{
site_quarto_build(
  quarto_base_folder = here::here(),
  verbosity = c("summary", "verbose", "silent")
)
}
\arguments{
\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.  Defaults to: 'verbose'}
}
\description{
Wrapper on the Quarto Render function
}
