% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone.R
\name{ecodown_clone}
\alias{ecodown_clone}
\title{Download the package's latest source code from repo}
\usage{
ecodown_clone(
  repo_url = "",
  commit = c("latest_tag", "latest_commit"),
  target_folder = tempdir(),
  branch = "main"
)
}
\arguments{
\item{repo_url}{Repo location}

\item{commit}{Commit to use as the base for the documentation.  It defaults
to 'latest_tag'. That default will search for the latest Git tag.  The
assumption is that the latest tag is the same as the latest release.  This
way we avoid documenting work-in-progress.  The 'latest_commit' value will
simply use whatever is cloned. Pass an SHA value if you wish to fix the
commit to use.}

\item{target_folder}{Location to copy the package to. Defaults to a temporary
directory}

\item{branch}{Repo branch. Defaults to 'main'}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.  Defaults to: 'verbose'}
}
\description{
Download the package's latest source code from repo
}
