% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{ecodown_quarto_render}
\alias{ecodown_quarto_render}
\title{Wrapper on the Quarto Render function}
\usage{
ecodown_quarto_render(
  quarto_folder = here::here(),
  verbosity = c("verbose", "summary", "silent"),
  autolink = FALSE,
  run = c("smart", "full", "no")
)
}
\arguments{
\item{quarto_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.}

\item{autolink}{Boolean flag that indicates if 'downlit' should run. Defaults
to TRUE.}

\item{run}{Flag that indicates the mode to render the site. Defaults to 'smart'
which attempts to run only the files that changed.}
}
\description{
Wrapper on the Quarto Render function
}
