% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{ecodown_quarto_render}
\alias{ecodown_quarto_render}
\title{Wrapper on the Quarto Render function}
\usage{
ecodown_quarto_render(
  quarto_folder = here::here(),
  verbosity = c("summary", "verbose", "silent")
)
}
\arguments{
\item{quarto_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.}
}
\description{
Wrapper on the Quarto Render function
}
