% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_articles}
\alias{package_articles}
\title{Copies the vignettes into Quarto}
\usage{
package_articles(
  pkg_folder = "",
  source = "vignettes",
  target = "articles",
  project_folder = "",
  root_folder = here::here()
)
}
\arguments{
\item{pkg_folder}{Path to the package's source code}

\item{source}{Sub-folder location where the vignettes are available}

\item{target}{Sub-folder location to place the articles in Quarto}

\item{project_folder}{Sub folder in \code{root_folder} that will be the base for
the package's documentation.}

\item{root_folder}{Base target Quarto folder. Defaults to current workspace.}
}
\description{
Copies the vignettes into Quarto
}
