% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_build_documentation}
\alias{package_build_documentation}
\title{Copies and/or converts files from package source into Quarto}
\usage{
package_build_documentation(
  package_source_folder = "",
  quarto_sub_folder = "",
  quarto_base_folder = here::here(),
  convert_readme = TRUE,
  convert_news = TRUE,
  convert_articles = TRUE,
  convert_reference = TRUE,
  downlit_options = TRUE,
  url_prefix = ""
)
}
\arguments{
\item{package_source_folder}{Path to the package's source code}

\item{quarto_sub_folder}{Sub folder in \code{quarto_base_folder} that will be the base for
the package's documentation.}

\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{convert_readme}{Flag that indicates if the README file needs to be processed}

\item{convert_news}{Flag that indicates if the NEWS file needs to be processed}

\item{convert_articles}{Flag that indicates if the vignette files needs to be processed}

\item{convert_reference}{Flag that indicates if the help files needs to be processed}

\item{downlit_options}{Flag that indicates if the package name should be
added to the 'options()' that tells 'downlit' that this is an internal
package}

\item{url_prefix}{String to prefix to the 'downlit' URL's}
}
\description{
Copies and/or converts files from package source into Quarto
}
