% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{package_clone_git_repo}
\alias{package_clone_git_repo}
\title{Download the package's latest source code from repo}
\usage{
package_clone_git_repo(url = "", target_folder = tempdir(), branch = "main")
}
\arguments{
\item{url}{Repo location}

\item{target_folder}{Location to copy the package to. Defaults to a temporary
directory}

\item{branch}{Repo branch. Defaults to 'main'}
}
\description{
Download the package's latest source code from repo
}
