% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_reference.R
\name{package_reference}
\alias{package_reference}
\alias{package_reference_index}
\alias{package_reference_pages}
\title{Processes reference/help files}
\usage{
package_reference(
  package_source_folder = "",
  quarto_base_folder = here::here(),
  quarto_sub_folder = "",
  reference_folder = "reference",
  downlit_options = TRUE,
  site_url = get_quarto_entry(quarto_base_folder, "site", "site-url")
)

package_reference_index(
  package_source_folder = "",
  reference_folder = "reference",
  quarto_sub_folder = "",
  quarto_base_folder = here::here(),
  pkg = NULL,
  downlit_options = TRUE,
  site_url = get_quarto_entry(quarto_base_folder, "site", "site-url")
)

package_reference_pages(
  package_source_folder = "",
  reference_folder = "reference",
  quarto_sub_folder = "",
  quarto_base_folder = here::here(),
  pkg = NULL,
  downlit_options = TRUE,
  site_url = get_quarto_entry(quarto_base_folder, "site", "site-url")
)
}
\arguments{
\item{package_source_folder}{Path to the package's source code}

\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{quarto_sub_folder}{Sub folder in \code{quarto_base_folder} that will be the base for
the package's documentation.}

\item{reference_folder}{Sub folder in \code{quarto_sub_folder} where the output files
will land.}

\item{downlit_options}{Flag that indicates if the package name should be
added to the 'options()' that tells 'downlit' that this is an internal
package}

\item{site_url}{URL of the target site.  It defaults to using the address in
the '_quarto.yml' file}

\item{pkg}{A \code{pkgdown} object. If one is passed, then the \code{package_source_folder} argument
will be ignored}
}
\description{
Processes reference/help files
}
