% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_autolink.R
\name{site_autolink_html}
\alias{site_autolink_html}
\title{Autolink function calls}
\usage{
site_autolink_html(
  quarto_base_folder = here::here(),
  render_folder = get_quarto_entry(quarto_base_folder, "project", "output-dir"),
  verbosity = c("summary", "verbose", "silent")
)
}
\arguments{
\item{quarto_base_folder}{Base target Quarto folder. Defaults to current workspace.}

\item{render_folder}{Location of the sub-folder that contains the output
from Quarto. It defaults to the 'output-dir' entry in the '_quarto.yml' file.}

\item{verbosity}{Level of messaging available during run time. Possible values
are 'verbose', 'summary', and 'silent'.  Defaults to: 'verbose'}
}
\description{
Autolink function calls
}
