% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{site_autolink_html}
\alias{site_autolink_html}
\title{Autolink function calls}
\usage{
site_autolink_html(quarto_rendered_site = "_site")
}
\arguments{
\item{quarto_rendered_site}{Location of the sub-folder that contains the output from
Quarto.}
}
\description{
Autolink function calls
}
